requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
Pillow
numpy